fix(website): guard SSR against window access for playground
Docusaurus executes client modules during the server-side build to
prerender pages. The playground page statically imported Monaco Editor
and the suppressResizeObserver client module, both of which touch
window/document at module init and blew up the SSR pass with
'ReferenceError: window is not defined'.

- Wrap the playground page body in BrowserOnly so Monaco, the codegen
  browser bundle, and their dependencies are only evaluated on the
  client. The content lives in a new PlaygroundContent component that is
  required lazily inside the BrowserOnly render function.
- Guard the suppressResizeObserver client module with typeof window so
  it's a no-op during SSR.
